+++ R/TuningInstanceAsyncSingleCrit.R
@@ -25,6 +25,7 @@
 #' @template param_measure
 #' @template param_terminator
 #' @template param_search_space
+#' @template param_internal_search_space
 #' @template param_store_benchmark_result
 #' @template param_store_models
 #' @template param_check_values
@@ -35,6 +36,7 @@
 #' @template param_xdt
 #' @template param_learner_param_vals
 #' @template param_internal_tuned_values
+#' @template param_extra
 #'
 #' @template field_internal_search_space
 #'
@@ -54,6 +56,7 @@ TuningInstanceAsyncSingleCrit = R6Class("TuningInstanceAsyncSingleCrit",
       measure = NULL,
       terminator,
       search_space = NULL,
+      internal_search_space = NULL,
       store_benchmark_result = TRUE,
       store_models = FALSE,
       check_values = FALSE,
@@ -67,24 +70,56 @@ TuningInstanceAsyncSingleCrit = R6Class("TuningInstanceAsyncSingleCrit",
       if (!is.null(search_space) && length(learner$param_set$get_values(type = "only_token"))) {
         stop("If the values of the ParamSet of the Learner contain TuneTokens you cannot supply a search_space.")
       }
-      if (is.null(search_space)) {
-        search_space = as_search_space(learner)
-        learner$param_set$values = learner$param_set$get_values(type = "without_token")
+
+      search_space_from_tokens = is.null(search_space)
+
+      # convert tune token to search space
+      search_space = if (is.null(search_space)) {
+        learner$param_set$search_space()
       } else {
-        search_space = as_search_space(search_space)
+        as_search_space(search_space)
       }
 
-      # internal search space
-      internal_tune_ids = keep(names(search_space$tags), map_lgl(search_space$tags, function(tag) "internal_tuning" %in% tag))
-      if (length(internal_tune_ids)) {
-        self$internal_search_space = search_space$subset(internal_tune_ids)
+      # get ids of primary and internal hyperparameters
+      sids = search_space$ids()
+      internal_tune_ids = search_space$ids(any_tags = "internal_tuning")
 
-        if (self$internal_search_space$has_trafo) {
-          stopf("Inner tuning and parameter transformations are currently not supported.")
+      # get internal search space
+      self$internal_search_space = if (is.null(internal_search_space)) {
+        # We DO NOT subset the search space because there we might keep an extra_trafo which is not allowed
+        # for the internal tuning search space
+        if (length(internal_tune_ids)) {
+          if (search_space_from_tokens) {
+            learner$param_set$subset(internal_tune_ids)$search_space()
+          } else {
+            search_space$subset(internal_tune_ids)
+          }
         }
+      } else {
+        if (length(internal_tune_ids)) {
+          stopf("Either tag parameters in the `search_space` with 'internal_tuning' OR provide an `internal_search_space`.")
+        }
+        as_search_space(internal_search_space)
+      }
 
-        search_space = search_space$subset(setdiff(search_space$ids(), internal_tune_ids))
+      # subset search space to primary hyperparameters
+      if (length(internal_tune_ids)) {
+        search_space = search_space$subset(setdiff(sids, internal_tune_ids))
+      }
 
+      # set learner parameter values
+      if (search_space_from_tokens) {
+        learner$param_set$values = learner$param_set$get_values(type = "without_token")
+      }
+
+      if (!is.null(self$internal_search_space) && self$internal_search_space$has_trafo) {
+        stopf("Internal tuning and parameter transformations are currently not supported.
+          If you manually provided a search space that has a trafo and parameters tagged with 'internal_tuning',
+          please pass the latter separately via the argument `internal_search_space`.")
+      }
+
+      # set internal search space
+      if (!is.null(self$internal_search_space)) {
         # the learner dictates how to interpret the to_tune(..., inner)
         learner$param_set$set_values(.values = learner$param_set$convert_internal_search_space(self$internal_search_space))
       }
@@ -128,11 +163,21 @@ TuningInstanceAsyncSingleCrit = R6Class("TuningInstanceAsyncSingleCrit",
     #'
     #' @param y (`numeric(1)`)\cr
     #' Optimal outcome.
+    #' @param xydt (`data.table::data.table()`)\cr
+    #' Point, outcome, and additional information (Deprecated).
     #' @param ... (`any`)\cr
     #' ignored.
-    assign_result = function(xdt, y, learner_param_vals = NULL, ...) {
+    assign_result = function(xdt, y, learner_param_vals = NULL, extra = NULL, xydt = NULL, ...) {
+      # workaround
+      extra = extra %??% xydt
+
       # set the column with the learner param_vals that were not optimized over but set implicitly
       assert_list(learner_param_vals, null.ok = TRUE, names = "named")
+
+      # extract internal tuned values
+      if ("internal_tuned_values" %in% names(extra)) {
+        set(xdt, j = "internal_tuned_values", value = list(extra[["internal_tuned_values"]]))
+      }
 
       if (is.null(learner_param_vals)) {
         learner_param_vals = self$objective$learner$param_set$values