+++ R/auto_tuner.R
@@ -16,13 +16,14 @@
 #' @template param_term_time
 #' @template param_terminator
 #' @template param_search_space
+#' @template param_internal_search_space
 #' @template param_store_tuning_instance
 #' @template param_store_benchmark_result
 #' @template param_store_models
 #' @template param_check_values
 #' @template param_callbacks
-#' @template param_validate
 #' @template param_rush
+#' @template param_id
 #'
 #' @export
 #' @examples
@@ -43,13 +44,14 @@ auto_tuner = function(
   term_time = NULL,
   terminator = NULL,
   search_space = NULL,
+  internal_search_space = NULL,
   store_tuning_instance = TRUE,
   store_benchmark_result = TRUE,
   store_models = FALSE,
   check_values = FALSE,
   callbacks = NULL,
-  validate = NULL,
-  rush = NULL
+  rush = NULL,
+  id = NULL
   ) {
   terminator = terminator %??% terminator_selection(term_evals, term_time)
 
@@ -60,11 +62,12 @@ auto_tuner = function(
     measure = measure,
     terminator = terminator,
     search_space = search_space,
+    internal_search_space = internal_search_space,
     store_tuning_instance = store_tuning_instance,
     store_benchmark_result = store_benchmark_result,
     store_models = store_models,
     check_values = check_values,
     callbacks = callbacks,
-    validate = validate,
-    rush = rush)
+    rush = rush,
+    id = id)
 }