+++ tests/testthat/test_AutoTuner.R
@@ -337,14 +337,16 @@ test_that("AutoTuner hash works #647 in mlr3", {
     resampling = rsmp("holdout"),
     measure = msr("classif.ce"),
     terminator = trm("evals", n_evals = 4),
-    tuner = tnr("grid_search", resolution = 3))
+    tuner = tnr("grid_search", resolution = 3),
+    id = "at_1")
 
   at_2 = AutoTuner$new(
     learner = lrn("classif.rpart", cp = to_tune(0.01, 0.1)),
     resampling = rsmp("holdout"),
     measure = msr("classif.ce"),
     terminator = trm("evals", n_evals = 4),
-    tuner = tnr("grid_search", resolution = 3))
+    tuner = tnr("grid_search", resolution = 3),
+    id = "at_2")
 
   expect_true(at_1$hash != at_2$hash)
 
@@ -621,7 +623,7 @@ test_that("AutoTuner works with async tuner", {
   skip_if_not_installed("rush")
   flush_redis()
 
-  rush_plan(n_workers = 2)
+  rush::rush_plan(n_workers = 2)
   at = auto_tuner(
     tuner = tnr("async_random_search"),
     learner = lrn("classif.rpart", cp = to_tune(1e-04, 1e-1, logscale = TRUE)),
@@ -657,35 +659,43 @@ test_that("AutoTuner works with internal tuning and validation", {
   # the AutoTuner's validate field controls the validation data for the final model fit,
   # because it was set to NULL, the full data was used for the final model fit
   expect_true(is.null(at$model$learner$state$internal_valid_task_ids))
+})
 
-  # we can also still do early stopping on the final model fit if we want to
+test_that("AutoTuner works when internal_search_space is passed separately", {
+  task = tsk("iris")
+  search_space = ps(x = p_dbl(0.2, 0.3))
+  internal_search_space = ps(iter = p_int(upper = 1000L, aggr = function(x) length(x)))
   at = auto_tuner(
     tuner = tnr("random_search", batch_size = 2),
-    learner = lrn("classif.debug", iter = 1000L, x = to_tune(0.2, 0.3), early_stopping = TRUE, validate = "test"),
+    learner = lrn("classif.debug", early_stopping = TRUE, validate = "test"),
     resampling = rsmp("cv", folds = 3),
+    search_space = search_space,
+    internal_search_space = internal_search_space,
     measure = msr("classif.ce"),
     term_evals = 4
   )
-  expect_error(at$train(task), "when a validation task is present")
-  at$validate = 0.2
   at$train(task)
+  expect_equal(at$model$learner$param_set$values$iter, 3)
+  expect_false(at$model$learner$param_set$values$early_stopping)
+})
 
-  # early stopping was not disabled
-  expect_equal(at$model$learner$param_set$values$iter, 1000)
-  expect_true(at$model$learner$param_set$values$early_stopping)
-
-  # ratios are handled correctly for the AutoTuner
+test_that("AutoTuner works when internal_search_space is part of primary search space", {
+  task = tsk("iris")
+  search_space = ps(
+    x = p_dbl(0.2, 0.3),
+    iter = p_int(upper = 1000L, aggr = function(x) length(x), tags = "internal_tuning")
+    )
   at = auto_tuner(
-    tuner = tnr("random_search"),
-    learner = lrn("classif.debug", early_stopping = TRUE, x = to_tune(0, 1)),
-    resampling = rsmp("holdout"),
-    store_models = TRUE,
-    term_evals = 1
+    tuner = tnr("random_search", batch_size = 2),
+    learner = lrn("classif.debug", early_stopping = TRUE, validate = "test"),
+    resampling = rsmp("cv", folds = 3),
+    search_space = search_space,
+    measure = msr("classif.ce"),
+    term_evals = 4
   )
-  set_validate(at, final_validate = 0.2, validate = 0.2)
   at$train(task)
-  expect_true(is.null(task$internal_valid_task))
-  expect_equal(at$model$learner$validate, "predefined")
+  expect_equal(at$model$learner$param_set$values$iter, 3)
+  expect_false(at$model$learner$param_set$values$early_stopping)
 })
 
 test_that("AutoTuner works with set_validate function", {
@@ -696,52 +706,8 @@ test_that("AutoTuner works with set_validate function", {
     measure = msr("classif.ce"),
     term_evals = 4
   )
-  set_validate(at, validate = "test", final_validate = 0.3)
-  expect_equal(at$validate, 0.3)
+  set_validate(at, validate = "test")
   expect_equal(at$learner$validate, "test")
   set_validate(at, validate = NULL)
-  expect_equal(at$validate, 0.3)
+  expect_true(is.null(at$learner$validate))
 })
-
-test_that("AutoTuner validate field can only be set if the tuned learner supports validation", {
-  at = auto_tuner(
-    tuner = tnr("random_search", batch_size = 2),
-    learner = lrn("classif.debug", iter = 1000L, x = to_tune(0.2, 0.3), early_stopping = TRUE),
-    resampling = rsmp("cv", folds = 3),
-    measure = msr("classif.ce"),
-    term_evals = 4,
-    validate = 0.3
-  )
-  expect_equal(at$validate, 0.3)
-
-  expect_error(
-    auto_tuner(
-      tuner = tnr("random_search", batch_size = 2),
-      learner = lrn("classif.rpart"),
-      resampling = rsmp("cv", folds = 3),
-      measure = msr("classif.ce"),
-      term_evals = 4,
-      validate = 0.3
-    ),
-    "The learner that is tuned by"
-  )
-})
-
-test_that("AutoTuner validation set can be used for final model fit", {
-  at = auto_tuner(
-    tuner = tnr("random_search"),
-    learner = lrn("classif.debug", early_stopping = TRUE, x = to_tune(0, 1)),
-    resampling = rsmp("holdout"),
-    store_models = TRUE,
-    term_evals = 1
-  )
-
-  task = tsk("iris")
-  task$divide(ratio = 0.2)
-
-  set_validate(at, final_validate = "predefined", validate = "predefined")
-  at$train(task)
-  expect_equal(at$state$internal_valid_task_hash, task$internal_valid_task$hash)
-  expect_equal(at$state$validate, "predefined")
-})
-