+++ NEWS.md
@@ -1,8 +1,24 @@
 # mlr3tuning (development version)
 
-* refactor: Replace internal tuning callback.
-* fix: Delete intermediate `BenchmarkResult` in `ObjectiveTuningBatch` after optimization.
+* compatibility: mlr3 0.22.0
 * compatibility: Work with new irace 4.0.0
+
+# mlr3tuning 1.1.0
+
+* fix: The `as_data_table()` functions do not unnest the `x_domain` colum anymore by default.
+* fix: `to_tune(internal = TRUE)` now also works if non-internal tuning parameters require have an `.extra_trafo`.
+* feat: It is now possible to pass an `internal_search_space` manually.
+  This allows to use parameter transformations on the primary search space in combination with internal hyperparameter tuning.
+* refactor: The `Tuner` pass extra information of the result in the `extra` parameter now.
+
+# mlr3tuning 1.0.2
+
+* refactor: Extract internal tuned values in instance.
+
+# mlr3tuning 1.0.1
+
+* refactor: Replace internal tuning callback.
+* perf: Delete intermediate `BenchmarkResult` in `ObjectiveTuningBatch` after optimization.
 
 # mlr3tuning 1.0.0